Reorganise suggestion test package
