+++ src/test/java/com/notably/logic/commands/suggestion/ExitSuggestionCommandTest.java
@@ -1,4 +1,4 @@
-package com.notably.logic.suggestion.commands;
+package com.notably.logic.commands.suggestion;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,7 +7,6 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import com.notably.logic.commands.suggestion.HelpSuggestionCommand;
 import com.notably.model.Model;
 import com.notably.model.ModelManager;
 import com.notably.model.block.BlockModel;
@@ -17,7 +16,7 @@ import com.notably.model.suggestion.SuggestionModelImpl;
 import com.notably.model.viewstate.ViewStateModel;
 import com.notably.model.viewstate.ViewStateModelImpl;
 
-public class HelpSuggestionCommandTest {
+public class ExitSuggestionCommandTest {
     private static Model model;
 
     @BeforeAll
@@ -32,12 +31,12 @@ public class HelpSuggestionCommandTest {
     @Test
     public void execute() {
         // Instantiate command
-        HelpSuggestionCommand helpSuggestionCommand = new HelpSuggestionCommand();
+        ExitSuggestionCommand exitSuggestionCommand = new ExitSuggestionCommand();
 
-        helpSuggestionCommand.execute(model);
+        exitSuggestionCommand.execute(model);
 
         // Expected result
-        String expectedResponseText = "Display a list of available commands";
+        String expectedResponseText = "Exit Notably app";
 
         assertEquals(Optional.of(expectedResponseText), model.responseTextProperty().getValue());
     }
